+++ Cargo.toml
@@ -54,7 +54,6 @@ members = [
     "core/schema-checker/schema-checker-core",
     "core/schema-checker/schema-checker-macro",
     "core/schema-checker/schema-checker-lib",
-    "core/thread-pools",
     "core/time",
     "genesis-tools/genesis-csv-to-json",
     "genesis-tools/genesis-populate",
@@ -285,7 +284,6 @@ near-schema-checker-macro = { path = "core/schema-checker/schema-checker-macro" 
 near-schema-checker-core = { path = "core/schema-checker/schema-checker-core" }
 near-schema-checker-lib = { path = "core/schema-checker/schema-checker-lib" }
 near-store = { path = "core/store" }
-near-thread-pools = { path = "core/thread-pools" }
 near-rpc-probe = { path = "benchmarks/rpc-probe" }
 near-transactions-generator = { path = "benchmarks/transactions-generator" }
 near-telemetry = { path = "chain/telemetry" }
