+++ chain/chain/Cargo.toml
@@ -34,7 +34,6 @@ reed-solomon-erasure.workspace = true
 
 near-async.workspace = true
 near-cache.workspace = true
-near-thread-pools.workspace = true
 near-chain-configs = { workspace = true, features = [
     "test_genesis",
     "test_utils",
