--- core/thread-pools/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "near-thread-pools"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-description = "Centralized thread pool definitions for nearcore"
-repository.workspace = true
-license.workspace = true
-publish = false
-
-[lints]
-workspace = true
-
-[dependencies]
-near-async.workspace = true
